requirements.txt
Flask==2.2.2
pymupdf==1.23.20
pytesseract==0.3.10
Pillow==9.4.0